refactor(checklist): extract shared shell and wealth-clause flag
- Compose DeductionCard and GrossIncomeCard from ChecklistCardShell; move optional
  amount inputs into ChecklistInlineAmountFields (parallel to the shell).
- Remove disclaimer badges; replace disclaimer_level with show_wealth_clause_notice.
- Align export Markdown header and results-page usage reminder with the new UX
  (shared copy in checklistCardCopy).
- Move gross-income salary footnotes into item.limitations; update types, content,
  tests, and ai-context docs.
